Ep7 Taking Damage

+++ Assets/Scripts/PlayerWeapon.cs
@@ -5,7 +5,7 @@
     public class PlayerWeapon
     {
         public string Name = "Glock";
-        public float Damage = 10f;
+        public int Damage = 10;
         public float Range = 100f;
